+++ src/commands/snapshot.zsh
@@ -1,18 +1,18 @@
 antigen-snapshot () {
   local snapshot_file="${1:-antigen-shapshot}"
+  local urls url dir version_hash snapshot_content
+  local -a bundles
 
   # The snapshot content lines are pairs of repo-url and git version hash, in
   # the form:
   #   <version-hash> <repo-url>
-  local snapshot_content="$(
-    -antigen-echo-record |
-    awk '$4 == "true" {print $1}' |
-    sort -u |
-    while read url; do
-      local dir="$(-antigen-get-clone-dir "$url")"
-      local version_hash="$(cd "$dir" && git rev-parse HEAD)"
-      echo "$version_hash $url"
-    done)"
+  urls=$(-antigen-echo-record | awk '$4 == "true" {print $1}' | sort -u)
+  for url in ${(f)urls}; do
+    dir="$(-antigen-get-clone-dir "$url")"
+    version_hash="$(\cd -q "$dir" && git rev-parse HEAD)"
+    bundles+=("$version_hash $url");
+  done
+  snapshot_content=${(j:\n:)bundles}
 
   {
     # The first line in the snapshot file is for metadata, in the form:
@@ -37,4 +37,3 @@ antigen-snapshot () {
 
   } > "$snapshot_file"
 }
-