+++ src/lib/revert-info.zsh
@@ -8,6 +8,7 @@
 # Returns
 #    Nothing. Generates/updates $ADOTDIR/revert-info.
 -antigen-revert-info() {
+  local url
   # Update your bundles, i.e., `git pull` in all the plugin repos.
   date >! $ADOTDIR/revert-info
 
@@ -15,9 +16,8 @@
     local clone_dir="$(-antigen-get-clone-dir "$url")"
     if [[ -d "$clone_dir" ]]; then
       (echo -n "$clone_dir:"
-        cd "$clone_dir"
+        \cd -q "$clone_dir"
         git rev-parse HEAD) >> $ADOTDIR/revert-info
     fi
   done
 }
-