+++ src/lib/use-prezto.zsh
@@ -1,4 +1,4 @@
 -antigen-use-prezto () {
+  ANTIGEN_DEFAULT_REPO_URL=$ANTIGEN_PREZTO_REPO_URL
   antigen-bundle "$ANTIGEN_PREZTO_REPO_URL"
 }
-